v0.0.1

+++ Notification/src/main/java/com/example/notification/repository/UserRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<UserEntity, Integer> {
-    Optional<UserEntity> findByUserIdAndDeliveryMethod(int userId, DeliveryMethod deliveryMethod);
+    Optional<UserEntity> findByUserIdAndDeliveryMethod(long userId, DeliveryMethod deliveryMethod);
 }
